// src/app/api/models/room-list-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RoomResponseDto } from '../models/room-response-dto';
export interface RoomListResponseDto {

  /**
   * Items per page
   */
  limit: number;

  /**
   * Current page number
   */
  page: number;
  rooms: Array<RoomResponseDto>;

  /**
   * Total number of rooms
   */
  total: number;

  /**
   * Total number of pages
   */
  totalPages: number;
}